+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * Allow other major versions of MiniTest than 4
 * Allow other verbs than 'have'
+* Add add_matchers method